Fixed Dijkstra map generating
Was looking at the camefrom table instead of the costsofar table...
clearly that makes a huge difference. Added a fancy function to write
the Dijkstra map to a txt file, and moved my logging functions to a more
appropriate log.h/c. Added a screenshot of the Dijkstra output to the
readme.